requirements.txt
asgiref==3.4.1
attrs==19.3.0
Automat==0.8.0
blinker==1.4
cachetools==4.2.4
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
Click==7.0
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
dbus-python==1.2.16
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.13
django-cors-headers==3.10.1
django-filter==21.1
django-model-utils==4.0.0
django-rest-knox==4.1.0
django-templated-mail==1.1.1
django-x509==0.9.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
entrypoints==0.3
facebook-sdk==3.1.0
future==0.18.2
google-api-core==2.2.2
google-api-python-client==2.32.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.20.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.8.2
incremental==16.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.10.1
jsonfield==3.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.3.6
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
oauth2client==4.1.3
oauthlib==3.1.0
openwisp-utils==0.7.5
packaging==21.3
pbr==5.8.0
PDFNetPython3==9.1.0
pexpect==4.6.0
Pillow==8.4.0
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
PyDrive==1.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.6
pyparted==3.11.2
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-debian===0.1.36ubuntu1
python-twitter==3.5
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sos==4.1
sqlparse==0.4.2
ssh-import-id==5.10
stripe==2.63.0
swapper==1.1.2.post1
systemd-python==234
testresources==2.0.1
Twisted==18.9.0
ubuntu-advantage-tools==27.4
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
wadllib==1.3.3
WALinuxAgent==2.2.46
zipp==1.0.0
zope.interface==4.7.1
